Вынес получение хоста в отдельный модуль, переработал логику sitemap
