+++ public/sitemap.xml
@@ -6,7 +6,4 @@
   <sitemap>
     <loc>https://razrabs.ru/feeds/sitemap.xml</loc>
   </sitemap>
-  <sitemap>
-    <loc>https://razrabs.ru/posts/sitemap.xml</loc>
-  </sitemap>
 </sitemapindex>
