+++ src/shared/lib/host/getHost.ts
@@ -1,0 +1,9 @@
+import getConfig from 'next/config'
+
+const { publicRuntimeConfig, serverRuntimeConfig } = getConfig()
+
+export const getHost = (): string => {
+  const isServerSide = typeof window === 'undefined'
+  const host = (isServerSide ? serverRuntimeConfig : publicRuntimeConfig).HOST
+  return host
+}
