+++ src/shared/lib/host/index.ts
@@ -1,0 +1,1 @@
+export { getHost } from './getHost'
