+++ src/shared/lib/index.ts
@@ -2,4 +2,5 @@ export * from './client-hints'
 export * from './helmet'
 export * from './hooks'
 export * from './pluralize'
+export * from './host'
 export { useDisplayAnimation } from './display-animation'
